SN32: Implement ST freerunning mode using CT+SysTick
Testing on an actual device indicated that the MR0 register of the
CT16B0 timer on SN32F240B is double-buffered (even though the datasheet
does not specify that) - the write to that register takes effect only
after the next wraparound of the timer, and there is no way to disable
that behavior and make the MR0 write take effect immediately.  This
makes the MR0 register unusable for the ST freerunning mode (the
interrupt on match does not come at the expected time).

Reimplement the ST freerunning mode by using the SysTick timer to
generate the alarm interrupts.  This requires some calculations in
st_lld_start_alarm() and st_lld_set_alarm(), but should make it possible
to have more precise delays than in the periodic tick mode.  The CT16Bx
timer is still used to get a stable time source (the SysTick timer needs
to be restarted on every alarm setting, which would result in losing
some ticks if that timer is also used for long-term timekeeping).