RP2040: Init buffer position in adc_lld_start_conversion()
The RP2040 ADC driver was initializing adcp->current_buffer_position and
friends only in adc_lld_start(), therefore a simple adcConvert() call
did not work when called a second time (only the circular mode worked,
because the interrupt handler was resetting the position in that case
after the whole buffer was filled).  Reset the buffer position variables
in adc_lld_start_conversion() to make the one-time conversion work.